+++ CHANGELOG.md
@@ -158,6 +158,8 @@
 
 * Cloud Access
     * Fixed issue in the `akamai_cloudaccess_key` resource that using variables to pass credentials could break the processing ([I#750](https://github.com/akamai/terraform-provider-akamai/issues/750)).
+* Appsec
+  * Fixed issue in the `akamai_appsec_rapid_rules` resource where `plan` and `apply` were showing diff due to a change in the rule ordering.
 
 * IAM
   * Fixed a value conversion error in the `akamai_iam_api_client` resource that occurred when `api_access.apis` elements were unknown at plan ([I#744](https://github.com/akamai/terraform-provider-akamai/issues/744)).